增加多线程下载

+++ app/src/main/java/com/hnpolice/xiaoke/downloadfile/bean/ThreadInfo.java
@@ -7,14 +7,14 @@ package com.hnpolice.xiaoke.downloadfile.bean;
 public class ThreadInfo {
     private int id;
     private String url;
-    private int start;
+    private long start;
     private long end;
     private long finish;
 
     public ThreadInfo() {
     }
 
-    public ThreadInfo(int id, String url, int start, long end, long finish) {
+    public ThreadInfo(int id, String url, long start, long end, long finish) {
         this.id = id;
         this.url = url;
         this.start = start;
@@ -38,11 +38,11 @@ public class ThreadInfo {
         this.url = url;
     }
 
-    public int getStart() {
+    public long getStart() {
         return start;
     }
 
-    public void setStart(int start) {
+    public void setStart(long start) {
         this.start = start;
     }
